refactor/EV-734/Adding Defaults loaders & screen errors
