+++ lib/modules/details/repositories/market_chart_repository_provider.dart
@@ -1,6 +1,5 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import '../api/crypto_base_endpoint_provider.dart';
+import '../../../shared/api/crypto_base_endpoint_provider.dart';
 import 'market_chart_repository.dart';
 
 final marketChartRepositoryProvider = Provider(
