+++ lib/modules/details/useCase/get_crypto_market_chart_useCase.dart
@@ -1,7 +1,7 @@
 import 'package:everest_flutter_crypto_list/modules/details/mapper/market_chart_mapper.dart';
 
-import '../../../shared/repositories/market_chart_repository.dart';
 import '../model/market_chart_view_data.dart';
+import '../repositories/market_chart_repository.dart';
 
 class GetCryptoMarketChartUseCase {
   final MarketChartRepository repository;
