+++ lib/modules/details/widgets/details_body.dart
@@ -1,6 +1,8 @@
+import 'package:everest_flutter_crypto_list/shared/widgets/default_loading_spinner.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
+import '../../../shared/widgets/default_error_widget.dart';
 import '../model/crypto_data_arguments.dart';
 import '../providers/providers.dart';
 import 'convert_crypto_button.dart';
@@ -31,16 +33,16 @@ class DetailsBody extends HookConsumerWidget {
               CryptoInformation(
                   marketChartData: data,
                   cryptoDataArguments: cryptoDataArguments),
-              ConvertCryptoButton(
-                  cryptoDataArguments: cryptoDataArguments),
+              ConvertCryptoButton(cryptoDataArguments: cryptoDataArguments),
             ],
           ),
         );
       },
-      error: (error, stackTrace) => Center(child: Text('$error')),
-      loading: () => const Center(
-        child: CircularProgressIndicator(),
-      ),
+      error: (error, stackTrace) {
+        debugPrintStack(stackTrace: stackTrace);
+        return const DefaultErrorWidget();
+      },
+      loading: () => const DefaultLoadingSpinner(),
     );
   }
 }
