+++ lib/modules/wallet/providers/providers.dart
@@ -1,8 +1,8 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../shared/repositories/crypto_data_repository_provider.dart';
 import '../model/crypto_list_view_data.dart';
 import '../model/user_wallet_model.dart';
+import '../repositories/crypto_data_repository_provider.dart';
 import '../useCases/get_cryptos_data_useCase.dart';
 
 final visibilityProvider = StateProvider<bool>(
