+++ lib/modules/wallet/repositories/crypto_data_repository.dart
@@ -1,6 +1,6 @@
 import 'package:everest_flutter_crypto_list/shared/api/crypto_base_endpoint.dart';
 
-import '../api/model/crypto_data_response.dart';
+import '../../../shared/api/model/crypto_data_response.dart';
 
 class CryptoDataRepository {
   final CryptoBaseEndpoint baseEndpoint;
