+++ lib/modules/wallet/repositories/crypto_data_repository_provider.dart
@@ -1,6 +1,6 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../api/crypto_base_endpoint_provider.dart';
+import '../../../shared/api/crypto_base_endpoint_provider.dart';
 import 'crypto_data_repository.dart';
 
 final cryptoDataRepositoryProvider = Provider(
