+++ lib/modules/wallet/views/wallet_screen.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import '../../../shared/widgets/default_error_widget.dart';
+import '../../../shared/widgets/default_loading_spinner.dart';
 import '../model/user_wallet_model.dart';
-import '../../../shared/providers/providers.dart';
 import '../providers/providers.dart';
 import '../widgets/crypto_list_view.dart';
 import '../widgets/header.dart';
@@ -33,13 +34,11 @@ class WalletScreen extends HookConsumerWidget {
           ),
         );
       },
-      error: (error, stackTrace) =>
-          Text("The following error was found on UserWallet: $error"),
-      loading: () {
-        return const Center(
-          child: CircularProgressIndicator(),
-        );
+      error: (error, stackTrace) {
+        debugPrintStack(stackTrace: stackTrace);
+        return const DefaultErrorWidget();
       },
+      loading: () => const DefaultLoadingSpinner(),
     );
   }
 }
