+++ pubspec.yaml
@@ -24,6 +24,8 @@ dependencies:
   brasil_fields: ^1.5.0
   currency_text_input_formatter: ^2.1.7
   mask_text_input_formatter: ^2.4.0
+  lottie: ^1.4.3
+  flutter_spinkit: ^5.1.0
 
 dev_dependencies:
   flutter_test:
